+++ resources/views/examples/knob/default.blade.php
@@ -1,0 +1,1 @@
+<x-ui.knob :value="50" />