+++ resources/views/examples/knob/sizes.blade.php
@@ -1,0 +1,5 @@
+<div class="flex flex-wrap items-center gap-8">
+    <x-ui.knob :value="25" size="sm" label="Small" />
+    <x-ui.knob :value="50" label="Default" />
+    <x-ui.knob :value="75" size="lg" label="Large" />
+</div>