+++ resources/views/examples/presence/pulse.blade.php
@@ -1,0 +1,1 @@
+<x-ui.presence status="online" :pulse="true" show-label />