+++ resources/views/examples/segmented-control/default.blade.php
@@ -1,0 +1,9 @@
+<x-ui.segmented-control
+    name="view"
+    value="list"
+    :options="[
+        ['value' => 'list', 'label' => 'List'],
+        ['value' => 'grid', 'label' => 'Grid'],
+        ['value' => 'board', 'label' => 'Board'],
+    ]"
+/>