+++ resources/views/examples/segmented-control/with-icons.blade.php
@@ -1,0 +1,9 @@
+<x-ui.segmented-control
+    name="layout"
+    value="grid"
+    :options="[
+        ['value' => 'list', 'label' => 'List', 'icon' => 'list'],
+        ['value' => 'grid', 'label' => 'Grid', 'icon' => 'layout-grid'],
+        ['value' => 'board', 'label' => 'Board', 'icon' => 'columns-3'],
+    ]"
+/>