+++ resources/views/examples/signature-pad/default.blade.php
@@ -1,0 +1,1 @@
+<x-ui.signature-pad name="signature" />