+++ resources/views/examples/signature-pad/with-label.blade.php
@@ -1,0 +1,4 @@
+<div class="flex w-full flex-col gap-2">
+    <x-ui.label for="sig">Signature</x-ui.label>
+    <x-ui.signature-pad id="sig" name="signature" />
+</div>